Remove bad assert nb_gpus > 1
This made benchmarker to fail when called with
`--framework=tensorflow` and multiple GPUs e.g. `--gpus=0,1`.  The
reason for this was, that the `super().__init__(params, ...)` was not
executed, ergo the `Benchmark` object didn't have a `params`
member... which is very bad.

On the other hand, I think this is handled by the "strategy" part, so
it can be removed.