Gpus issue gh48 (#54)
* Compare tf gpus with param gpus

* Fix CUDA_VISIBLE_DEVICES

* Remove bad assert nb_gpus > 1

This made benchmarker to fail when called with
`--framework=tensorflow` and multiple GPUs e.g. `--gpus=0,1`.  The
reason for this was, that the `super().__init__(params, ...)` was not
executed, ergo the `Benchmark` object didn't have a `params`
member... which is very bad.

On the other hand, I think this is handled by the "strategy" part, so
it can be removed.

* Clean up, add nice error message

* Remove unreachable code

The assertion above from one of the previous comment, ensures the GPU
specified in the params, and the GPUs "seen" by TF are the same.